Boot to Finder desktop instead of nanokernel test
Disable nanokernel threading test in favor of normal desktop boot.
The nanokernel is still fully functional - it's just not running test
threads on boot. System now proceeds to Finder event loop and displays
the classic System 7 desktop with menu bar and volume icons.

Testing can be re-enabled by changing #if 0 to #if 1 in main.c:2094.